Added first layer of frontend

+++ dynamik_frontend/src/components/DevsList.js
@@ -1,32 +1,40 @@
 import React, { useState } from "react";
 
-export default function PersonList({ onDeleteDev, onDevsList }) {
+export default function DevList({ onDeleteDev, onDevsList, onSelectDev }) {
   const [details, setDetails] = useState(null);
 
   const onChosenDev = (dev) => {
-    setDetails(dev);
+    setDetails((currentDetails) => (currentDetails === dev ? null : dev));
   };
 
   const handleDeleteDev = (dev) => {
-    onDeleteDev(dev._id);
+    if (window.confirm("Are you sure?")) onDeleteDev(dev._id);
   };
 
+  function Dev({ dev }) {
+    return (
+      <li onClick={() => onChosenDev(dev)}>
+        <h3>{dev.name}</h3>
+        <div>
+          <p>
+            <span>Birth Date: {dev.birth_date}</span>
+          </p>
+          <p>
+            <span>Stack: {dev.stack}</span>
+          </p>
+        </div>
+        <button className="btn-delete" onClick={() => handleDeleteDev(dev)}>
+          X
+        </button>
+      </li>
+    );
+  }
+
   return (
-    <ul>
-      {onDevsList.map((dev) => (
-        <li key={dev._id}>
-          {dev.name}
-          <button onClick={() => onChosenDev(dev)}>Show Details</button>
-          <button onClick={() => handleDeleteDev(dev)}>❌</button>
-        </li>
+    <ul className="list list-devs">
+      {onDevsList?.map((dev) => (
+        <Dev dev={dev} key={dev._id} />
       ))}
-      {details && (
-        <ul>
-          <li>{details.nickname}</li>
-          <li>{details.birth_date}</li>
-          <li>{details.stack}</li>
-        </ul>
-      )}
     </ul>
   );
 }
